Add 'check-end-of-life.sh' build script
Besides using the `endoflife.date` API to check support status, we also test whether the Nextcloud version to build is older than a previously built Nextcloud version, because Nextcloud can't be downgraded. This can happen when Nextcloud withdraws a version due to a regression (e.g. Nextcloud 29.0.9 was withdrawn due to a regression in the PDF viewer app). This scrip only runs on auto updates, i.e. manually building an older version of Nextcloud is still possible.